+++ CHANGELOG.md
@@ -53,6 +53,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Frame3D: import/export step.
 - BSplineFace3D: neutral_fiber.
 - Edge: from_step trim of periodic curves with different orientation of original edge
+- Arc3D: fix abscissa, fix get_arc_point_angle
 
 ### Refactor
 - ClosedShell3D: point_belongs, get_non_intersecting_faces